geo_ai_backend/pyproject.toml
[tool.poetry]
name = "GEO_AI_BACKEND"
version = "0.1.0"
description = ""
authors = []



[tool.poetry.dependencies]
python = ">=3.9, <3.10"
fastapi = "0.70.1"
uvicorn = {extras = ["standard"], version = "0.17.6"}
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = {extras = ["email"], version = "^1.10.7"}
requests = "^2.30.0"
pre-commit = "^3.3.1"
flake8 = "^6.0.0"
fastapi-sqlalchemy = "^0.2.1"
alembic = "^1.11.0"
# psycopg2 = "^2.9.6"
python-dotenv = "^1.0.0"
bcrypt = "^4.0.1"
pytest = "^7.4.0"
types-requests = "^2.31.0.2"
psycopg2-binary = "^2.9.7"
celery = "^5.3.1"
redis = "^4.6.0"
numpy = "1.24.2"
opencv-python = "4.7.0.72"
tritonclient = "^2.36.0"
gevent = "23.7.0"
geventhttpclient = "^2.0.9"
pillow = "^10.0.0"
ldap3 = "^2.9.1"
asyncio = "^3.4.3"
torchvision = "0.15.2"
torch = "2.0.1"
nvidia-cublas-cu11 = "11.10.3.66"
nvidia-cuda-cupti-cu11 = "11.7.101"
nvidia-cuda-nvrtc-cu11 = "11.7.99"
nvidia-cuda-runtime-cu11 = "11.7.99"
nvidia-cufft-cu11 = "10.9.0.58"
nvidia-curand-cu11 = "10.2.10.91"
nvidia-cusolver-cu11 = "11.4.0.1"
nvidia-cusparse-cu11 = "11.7.4.91"
nvidia-nvtx-cu11 = "11.7.91"
albumentations = "1.3.0"
rasterio = "1.3.6"
pandas = "1.5.2"
fiona = "1.9.4.post1"
shapely = "2.0.1"
nvidia-cudnn-cu11 = "^8.9.4.25"
nvidia-nccl-cu11 = "2.14.3"
urllib3 = "1.26.14"
geopandas = "^0.13.2"
arcgis = "^2.1.0.3"
open3d = "0.16.0"
laspy = "^2.5.1"
#geo-ai-ocr = {path = "geo_ai_backend/ml/ml_models/geo_ai_ocr"}
onnxruntime-gpu = "^1.16.3"
easydict = "^1.11"
pyproj = "^3.6.1"
arabic-reshaper = "^3.0.0"
python-bidi = "^0.4.2"
easyocr = "^1.7.1"
fastapi-utils = "^0.2.1"

[tool.poetry.dev-dependencies]
pytest = ">=5.2"
isort = "^5.12.0"
black = "^23.3.0"
mypy = "^1.2.0"
commitizen = "^2.42.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = 'black'
